requirements.txt
flask
flask-cors
transformers
faiss-cpu
torch
numpy
langchain
